# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/staff": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "summary": "Get list of Staff",
                "description": "Returns list of Staff",
                "operationId": "getStaffList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Store Staff information",
                "description": "Returns Staff data",
                "operationId": "storeStaff",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Staff user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "StartDate",
                        "in": "path",
                        "description": "Staff StartDate",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "branche_id",
                        "in": "path",
                        "description": "Staff branche_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "role",
                        "in": "path",
                        "description": "Staff role",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/staff/{id}": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "summary": "Get Staff information",
                "description": "Returns Staff data",
                "operationId": "getStaffById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Staff id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "Staff"
                ],
                "summary": "Update Staff information",
                "description": "Returns Staff data",
                "operationId": "updateStaff",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Staff user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "StartDate",
                        "in": "path",
                        "description": "Staff StartDate",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "branche_id",
                        "in": "path",
                        "description": "Staff branche_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "role",
                        "in": "path",
                        "description": "Staff role",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Staff"
                ],
                "summary": "delete Staff information",
                "description": "delete Staff data",
                "operationId": "deleteStaffById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Staff id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/staff/restore/{id}": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "summary": "restore deleted staff information",
                "description": "restore staff data",
                "operationId": "restoreStaffById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Staff id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
        "/staff/restore": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "summary": "restore all deleted staff information",
                "description": "restore all  staff data",
                "operationId": "restoreAllStaff",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
    },
...
}